# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/public/kb/articles": {
      "get": {
        "description": "Without `q`, published articles newest first. With `q`, the best matches first, each with a snippet showing where the words matched. Filter by category slug or tag either way.",
        "operationId": "KbPublicController_articles",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Words to look for. \"Quoted phrases\", `or`, and `-word` to exclude are understood.",
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "A category's slug",
            "schema": {
              "maxLength": 100,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 40,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page, 1 to 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "`nextCursor` from the previous page",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticlePage_Output"
                }
              }
            }
          },
          "400": {
            "description": "A filter or the cursor is invalid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Browse or search articles",
        "tags": [
          "Knowledge base (public)"
        ]
      }
    },
    "/api/v1/public/kb/articles/{slug}": {
      "get": {
        "operationId": "KbPublicController_article",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "maxLength": 100,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticle_Output"
                }
              }
            }
          },
          "404": {
            "description": "No published article with that slug",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Read an article",
        "tags": [
          "Knowledge base (public)"
        ]
      }
    },
...
  },
  "info": {
    "title": "DSD Unified Customer Support API",
    "description": "The API behind the DSD customer app and the DSD agent app.\n\nAll routes except `/health` and `/ready` live under `/api/v1`. Errors use RFC 9457 problem details (`application/problem+json`); every response carries an `X-Request-Id` header that matches the `requestId` in an error body and the server logs.\n\n## Signing in\n\nCustomers and staff are separate realms with separate routes, cookies and sessions: `/api/v1/auth/customer/*` and `/api/v1/customer/*` for customers, `/api/v1/auth/staff/*` and `/api/v1/staff/*` for staff. A session from one realm is refused (401) by every route of the other.\n\nSigning in sets an HttpOnly session cookie and a readable CSRF cookie, and returns the CSRF token in the body too (it is also in `me`). Every POST, PUT, PATCH and DELETE must come from a trusted origin, and once signed in must send that token in the `X-CSRF-Token` header.\n\nTo try it here, sign in with \"Try it out\" on a `login` route: the browser keeps the cookies, and this page copies the CSRF token into the header for you.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "customerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-dsd_customer_session",
        "description": "Set by the customer login route. HttpOnly, so scripts can't read it."
      },
      "staffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-dsd_staff_session",
        "description": "Set by the staff login route. HttpOnly, so scripts can't read it."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "The csrfToken from login or me. Required with the session cookie on every state-changing request."
      }
    },
    "schemas": {
...
      "PublicArticlePage_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "slug": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "summary": {
                  "type": "string"
                },
                "category": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "slug": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "slug",
                        "name"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "publishedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                },
                "snippet": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "text": {
                            "type": "string"
                          },
                          "highlighted": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "text",
                          "highlighted"
                        ],
                        "additionalProperties": false
                      }
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "description": "Where the words matched, when searching; null when browsing"
                }
              },
              "required": [
                "slug",
                "title",
                "summary",
                "category",
                "tags",
                "publishedAt",
                "snippet"
              ],
              "additionalProperties": false
            }
          },
          "nextCursor": {
            "description": "Pass as `cursor` for the next page; null on the last page",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
      "PublicArticle_Output": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "bodyMarkdown": {
            "type": "string",
            "description": "Sanitised markdown: render it without raw HTML, as the help centre does"
          },
          "category": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "slug": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "slug",
                  "name"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "slug",
          "title",
          "summary",
          "bodyMarkdown",
          "category",
          "tags",
          "publishedAt"
        ],
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Problem type URI; `about:blank` when the HTTP status says it all"
          },
          "title": {
            "type": "string",
            "description": "Short summary of the problem type"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "description": "Explanation specific to this occurrence",
            "type": "string"
          },
          "instance": {
            "description": "The request path that produced the problem",
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "description": "Matches the X-Request-Id response header and the server logs"
          },
          "errors": {
            "description": "Present on validation problems: one entry per invalid field",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "description": "Dot-separated path to the invalid field; empty for the whole body"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          },
          "allowedTransitions": {
            "description": "Present on `invalid-status-transition`: the statuses the ticket can move to now",
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "open",
                "pending_customer",
                "resolved",
                "closed"
              ]
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "requestId"
        ]
      }
    }
  }
}